refactor(ui): organize components per folder and forward root styling
