+++ packages/ui/src/components/EmptyState/EmptyState.stories.tsx
@@ -1,4 +1,4 @@
-import { FourThemeModes } from "../storybook";
+import { FourThemeModes } from "../../storybook";
 
 import { EmptyState } from "./EmptyState";
