+++ packages/ui/src/components/ErrorState/ErrorState.stories.tsx
@@ -1,6 +1,6 @@
 import { expect, userEvent, within } from "storybook/test";
 
-import { FourThemeModes } from "../storybook";
+import { FourThemeModes } from "../../storybook";
 
 import { ErrorState } from "./ErrorState";
