+++ packages/ui/src/components/IconButton/IconButton.stories.tsx
@@ -1,6 +1,6 @@
 import { expect, userEvent, within } from "storybook/test";
 
-import { FourThemeModes } from "../storybook";
+import { FourThemeModes } from "../../storybook";
 
 import { IconButton } from "./IconButton";
