+++ packages/ui/src/components/StatusPill/StatusPill.stories.tsx
@@ -1,4 +1,4 @@
-import { FourThemeModes } from "../storybook";
+import { FourThemeModes } from "../../storybook";
 
 import { StatusPill } from "./StatusPill";
