+++ packages/ui/src/index.ts
@@ -1,14 +1,29 @@
-export { EmptyState, type EmptyStateProps } from "./components/EmptyState";
-export { ErrorState, type ErrorStateProps } from "./components/ErrorState";
-export { Icon, type IconProps } from "./components/Icon";
-export { IconButton, type IconButtonProps } from "./components/IconButton";
-export { ProgressBar, type ProgressBarProps } from "./components/ProgressBar";
-export { SearchInput, type SearchInputProps } from "./components/SearchInput";
-export { Spinner, type SpinnerProps } from "./components/Spinner";
+export {
+	EmptyState,
+	type EmptyStateProps,
+} from "./components/EmptyState/EmptyState";
+export {
+	ErrorState,
+	type ErrorStateProps,
+} from "./components/ErrorState/ErrorState";
+export { Icon, type IconProps } from "./components/Icon/Icon";
+export {
+	IconButton,
+	type IconButtonProps,
+} from "./components/IconButton/IconButton";
+export {
+	ProgressBar,
+	type ProgressBarProps,
+} from "./components/ProgressBar/ProgressBar";
+export {
+	SearchInput,
+	type SearchInputProps,
+} from "./components/SearchInput/SearchInput";
+export { Spinner, type SpinnerProps } from "./components/Spinner/Spinner";
 export {
 	StatusPill,
 	type StatusPillProps,
 	type StatusPillTone,
-} from "./components/StatusPill";
+} from "./components/StatusPill/StatusPill";
 export type { CodiconName } from "./codicons";
 export { useVscodeTheme, type VscodeThemeKind } from "./useVscodeTheme";
